Refactor engine.py and player.py
We moved everything in engine.py and player.py.
Basically we moved a lot of functions from Board to Game.
The Board classes are now responsible for maintaining the physical state of
the board and the Game function runs the different rules of the game.
The players are only here to tell which move is played and to keep a
track of the states of the lofts.