Update alphabeta.py
Add class that adapt list_coeff_gain to the stage of the game

I added the class AlphaBetaBegin and AlphaBetaMidgame that play only the begin of the game or the midgame.
I added the class AleaAlphaBeta that play randomly the begin of the game (to prepare the board for AlphaBetaMidgame), or the begin and the midgame (to prepare the board for AlphaBeta which plays the endgame).
I added the class TestAmelioration, which is only used by the genetic algorithm (because TestAmelioration can take a stage in parameter).
I added the notion "STOP" to end the game without distributing the staying seeds.

We discovered strange behaviours in AlphaBeta. We think that it is linked to the fact that it returns "END" even if it isn't the end.